+++ CHANGELOG.md
@@ -1,6 +1,9 @@
-## [15.0.0] - 12/xx/2023
+## [15.0.0] - 01/xx/2024
 
 * upgrades to Flutter `3.16.5`
+* adds new feature **Conditional Form Fields**
+  * adds `conditions` property to `FastFormField`
+  * adds `FastCondition` class
 * adds new `FastSegmentendButton` that wraps Material [`SegmentedButton`](https://api.flutter.dev/flutter/material/SegmentedButton-class.html?gclid=CjwKCAiAs6-sBhBmEiwA1Nl8s5c8wRhB056CtAeuZP1-m_XUYi-jYGrVMh1lMnkwRmvKwlTelErtixoCAS4QAvD_BwE&gclsrc=aw.ds)
 * adds `showInputDecoration` property to `FastCheckbox`, `FastSwitch`, `FastChoiceChips`, `FastRangeSlider` and `FastSlider`
 * adds `canPop` and `onPopInvoked` properties to `FastForm`